remove api.lua
