+++ xmake/languages/cuda/load.lua
@@ -18,17 +18,60 @@
 -- @file        load.lua
 --
 
--- imports
-import("api")
+function apiset_apis()
+    local apis = {}
+    apis.values = {
+        -- target.add_xxx
+        "target.add_links"
+    ,   "target.add_syslinks"
+    ,   "target.add_cugencodes"
+    ,   "target.add_cuflags"
+    ,   "target.add_culdflags"
+    ,   "target.add_ldflags"
+    ,   "target.add_arflags"
+    ,   "target.add_shflags"
+    ,   "target.add_defines"
+    ,   "target.add_undefines"
+    ,   "target.add_rpathdirs"  -- @note do not translate path, it's usually an absolute path or contains $ORIGIN/@loader_path
+        -- option.add_xxx
+    ,   "option.add_links"
+    ,   "option.add_syslinks"
+    ,   "option.add_cugencodes"
+    ,   "option.add_cuflags"
+    ,   "option.add_culdflags"
+    ,   "option.add_ldflags"
+    ,   "option.add_arflags"
+    ,   "option.add_shflags"
+    ,   "option.add_defines"
+    ,   "option.add_undefines"
+    ,   "option.add_rpathdirs"
+        -- toolchain.add_xxx
+    ,   "toolchain.add_links"
+    ,   "toolchain.add_syslinks"
+    ,   "toolchain.add_cugencodes"
+    ,   "toolchain.add_cuflags"
+    ,   "toolchain.add_culdflags"
+    ,   "toolchain.add_ldflags"
+    ,   "toolchain.add_arflags"
+    ,   "toolchain.add_shflags"
+    ,   "toolchain.add_defines"
+    ,   "toolchain.add_undefines"
+    ,   "toolchain.add_rpathdirs"
+    }
+    apis.paths = {
+        -- target.add_xxx
+        "target.add_linkdirs"
+    ,   "target.add_includedirs"
+    ,   "target.add_sysincludedirs"
+        -- option.add_xxx
+    ,   "option.add_linkdirs"
+    ,   "option.add_sysincludedirs"
+    }
+    return apis
+end
 
--- load it
 function main()
-
-    -- init apis
-    _g.apis = api.apis()
-
-    -- ok
-    return _g
+    return {apis = _get_apis()}
 end
