+++ xmake/languages/msrc/load.lua
@@ -18,17 +18,29 @@
 -- @file        load.lua
 --
 
--- imports
-import("api")
+function _get_apis()
+    local apis = {}
+    apis.values = {
+        -- target.add_xxx
+        "target.add_mrcflags"
+        -- option.add_xxx
+    ,   "option.add_mrcflags"
+        -- toolchain.add_xxx
+    ,   "toolchain.add_mrcflags"
+    }
+    apis.paths = {
+        -- target.add_xxx
+        "target.add_includedirs"
+    ,   "target.add_sysincludedirs"
+        -- option.add_xxx
+    ,   "option.add_includedirs"
+    ,   "option.add_sysincludedirs"
+    }
+    return apis
+end
 
--- load it
 function main()
-
-    -- init apis
-    _g.apis = api.apis()
-
-    -- ok
-    return _g
+    return {apis = _get_apis()}
 end
