+++ xmake/languages/objc++/load.lua
@@ -18,17 +18,83 @@
 -- @file        load.lua
 --
 
--- imports
-import("api")
+function _get_apis()
+    local apis = {}
+    apis.values = {
+        -- target.add_xxx
+        "target.add_links"
+    ,   "target.add_syslinks"
+    ,   "target.add_mflags"
+    ,   "target.add_mxflags"
+    ,   "target.add_mxxflags"
+    ,   "target.add_ldflags"
+    ,   "target.add_arflags"
+    ,   "target.add_shflags"
+    ,   "target.add_defines"
+    ,   "target.add_undefines"
+    ,   "target.add_frameworks"
+    ,   "target.add_rpathdirs"  -- @note do not translate path, it's usually an absolute path or contains $ORIGIN/@loader_path
+        -- option.add_xxx
+    ,   "option.add_cincludes"
+    ,   "option.add_cxxincludes"
+    ,   "option.add_cfuncs"
+    ,   "option.add_cxxfuncs"
+    ,   "option.add_ctypes"
+    ,   "option.add_cxxtypes"
+    ,   "option.add_links"
+    ,   "option.add_syslinks"
+    ,   "option.add_mflags"
+    ,   "option.add_mxflags"
+    ,   "option.add_mxxflags"
+    ,   "option.add_ldflags"
+    ,   "option.add_arflags"
+    ,   "option.add_shflags"
+    ,   "option.add_defines"
+    ,   "option.add_undefines"
+    ,   "option.add_frameworks"
+    ,   "option.add_rpathdirs"
+        -- toolchain.add_xxx
+    ,   "toolchain.add_links"
+    ,   "toolchain.add_syslinks"
+    ,   "toolchain.add_mflags"
+    ,   "toolchain.add_mxflags"
+    ,   "toolchain.add_mxxflags"
+    ,   "toolchain.add_ldflags"
+    ,   "toolchain.add_arflags"
+    ,   "toolchain.add_shflags"
+    ,   "toolchain.add_defines"
+    ,   "toolchain.add_undefines"
+    ,   "toolchain.add_frameworks"
+    ,   "toolchain.add_rpathdirs"
+    ,   "toolchain.add_linkdirs"
+    ,   "toolchain.add_includedirs"
+    ,   "toolchain.add_sysincludedirs"
+    ,   "toolchain.add_frameworkdirs"
+    }
+    apis.paths = {
+        -- target.set_xxx
+        "target.set_headerdir"      -- TODO deprecated
+    ,   "target.set_config_header"
+    ,   "target.set_pcheader"
+    ,   "target.set_pcxxheader"
+        -- target.add_xxx
+    ,   "target.add_headers"        -- TODO deprecated
+    ,   "target.add_headerfiles"
+    ,   "target.add_linkdirs"
+    ,   "target.add_includedirs"
+    ,   "target.add_sysincludedirs"
+    ,   "target.add_frameworkdirs"
+        -- option.add_xxx
+    ,   "option.add_linkdirs"
+    ,   "option.add_includedirs"
+    ,   "option.add_sysincludedirs"
+    ,   "option.add_frameworkdirs"
+    }
+    return apis
+end
 
--- load it
 function main()
-
-    -- init apis
-    _g.apis = api.apis()
-
-    -- ok
-    return _g
+    return {apis = _get_apis()}
 end
